+++ changelog.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## Unreleased
+
 - Add inputMaybeValueDefault option to allow default input values other than null
 
 ## 1.0.0 - 2022-05-07